+++ src/coreComponents/physicsSolvers/fluidFlow/CompositionalMultiphaseStatistics.cpp
@@ -96,7 +96,7 @@ void CompositionalMultiphaseStatistics::registerDataOnMesh( Group & meshBodies )
     // if we have to report region statistics, we have to register them first here
     if( m_computeRegionStatistics )
     {
-      for( integer i = 0; i < regionNames.size(); ++i )
+      for( size_t i = 0; i < regionNames.size(); ++i )
       {
         ElementRegionBase & region = elemManager.getRegion( regionNames[i] );
 
@@ -198,7 +198,7 @@ void CompositionalMultiphaseStatistics::computeRegionStatistics( real64 const ti
 
   // Step 1: initialize the average/min/max quantities
   ElementRegionManager & elemManager = mesh.getElemManager();
-  for( integer i = 0; i < regionNames.size(); ++i )
+  for( size_t i = 0; i < regionNames.size(); ++i )
   {
     ElementRegionBase & region = elemManager.getRegion( regionNames[i] );
     RegionStatistics & regionStatistics = region.getReference< RegionStatistics >( viewKeyStruct::regionStatisticsString() );
@@ -354,7 +354,7 @@ void CompositionalMultiphaseStatistics::computeRegionStatistics( real64 const ti
   } );
 
   // Step 3: synchronize the results over the MPI ranks
-  for( integer i = 0; i < regionNames.size(); ++i )
+  for( size_t i = 0; i < regionNames.size(); ++i )
   {
     ElementRegionBase & region = elemManager.getRegion( regionNames[i] );
     RegionStatistics & regionStatistics = region.getReference< RegionStatistics >( viewKeyStruct::regionStatisticsString() );
